edit tf-config
